NMS-8085: Reverted adding dependencies
